+++ src/lib/secretsManager/providerService/ISecretProvider.ts
@@ -1,0 +1,10 @@
+import { SecretProviderType } from "../types";
+
+export interface ISecretProvider {
+  type: SecretProviderType;
+  id: string;
+
+  testConnection(): Promise<boolean>;
+  fetchSecret(secretName: string): Promise<string>;
+  listSecrets(): Promise<string[]>;
+}
