+++ src/lib/secretsManager/types.ts
@@ -17,12 +17,3 @@ export interface SecretProviderConfig {
   updatedAt: number;
   config: AWSSecretsManagerConfig;
 }
-
-export interface ISecretProvider {
-  type: SecretProviderType;
-  id: string;
-
-  testConnection(): Promise<boolean>;
-  fetchSecret(secretName: string): Promise<string>;
-  listSecrets(): Promise<string[]>;
-}
